+++ .env.example
@@ -1,2 +1,3 @@
 PORT
 ACCESS_SECRET
+DATABASE_URL
