+++ gateways/http/accounts/create_account.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 
 	"github.com/Vivi3008/apiTestGolang/domain/entities/account"
-	"github.com/Vivi3008/apiTestGolang/http/response"
+	"github.com/Vivi3008/apiTestGolang/gateways/http/response"
 )
 
 type AccountRequest struct {
